+++ EnemyAbilities/EnemyAbilities.cs
@@ -24,7 +24,7 @@ namespace EnemyAbilities
         public const string PluginGUID = PluginAuthor + "." + PluginName;
         public const string PluginAuthor = "Skeletogne";
         public const string PluginName = "EnemyAbilities";
-        public const string PluginVersion = "1.9.0";
+        public const string PluginVersion = "1.10.0";
 
         internal static bool RooInstalled => Chainloader.PluginInfos.ContainsKey("com.rune580.riskofoptions");
         public static EnemyAbilities Instance { get; private set; }
@@ -49,6 +49,7 @@ namespace EnemyAbilities
         {
             Instance = this;
             Log.Init(Logger);
+            Utils.Init();
             PluginConfig.Init(Config);
             foreach(var kvp in configEntries)
             {